--- src/main/java/uk/gov/hmcts/reform/prl/controllers/solicitorresponse/ResponseController.java
@@ -1,100 +1,0 @@
-package uk.gov.hmcts.reform.prl.controllers.solicitorresponse;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RestController;
-import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
-import uk.gov.hmcts.reform.prl.enums.citizen.ConfidentialityListEnum;
-import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
-import uk.gov.hmcts.reform.prl.services.DocumentLanguageService;
-import uk.gov.hmcts.reform.prl.services.UserService;
-import uk.gov.hmcts.reform.prl.services.document.DocumentGenService;
-import uk.gov.hmcts.reform.prl.services.tab.alltabs.AllTabServiceImpl;
-import uk.gov.hmcts.reform.prl.utils.CaseUtils;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
-
-@Slf4j
-@RestController
-@RequiredArgsConstructor(onConstructor = @__(@Autowired))
-@SecurityRequirement(name = "Bearer Authentication")
-public class ResponseController {
-
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    @Autowired
-    private UserService userService;
-
-    @Autowired
-    private AllTabServiceImpl allTabService;
-
-    @Autowired
-    private DocumentGenService documentGenService;
-
-    @Autowired
-    DocumentLanguageService documentLanguageService;
-
-    @PostMapping(path = "/keep-details-private-list", consumes = APPLICATION_JSON, produces = APPLICATION_JSON)
-    @Operation(description = "Callback to send FL401 application notification. ")
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Application Submitted."),
-        @ApiResponse(responseCode = "400", description = "Bad Request")})
-    public AboutToStartOrSubmitCallbackResponse retrieveKeepDetailsPrivateDetails(@RequestHeader("Authorization") @Parameter(hidden = true)
-                                                                 String authorisation,
-                                                                                  @RequestBody CallbackRequest callbackRequest) {
-        CaseData caseData = objectMapper.convertValue(
-            callbackRequest.getCaseDetails().getData(),
-            CaseData.class
-        );
-        StringBuilder selectedList = new StringBuilder();
-
-        selectedList.append("<ul>");
-        for (ConfidentialityListEnum confidentiality: caseData.getKeepContactDetailsPrivateOther()
-            .getConfidentialityList()) {
-            selectedList.append("<li>");
-            selectedList.append(confidentiality.getDisplayedValue());
-            selectedList.append("</li>");
-        }
-        selectedList.append("</ul>");
-
-        Map<String, Object> keepDetailsPrivateList = new HashMap<>();
-        keepDetailsPrivateList.put("confidentialListDetails", selectedList);
-
-        return AboutToStartOrSubmitCallbackResponse.builder()
-            .data(keepDetailsPrivateList)
-            .build();
-    }
-
-    @PostMapping(path = "/generate-c7response-draft-document", consumes = APPLICATION_JSON, produces = APPLICATION_JSON)
-    @Operation(description = "Callback to generate and store document")
-    @SecurityRequirement(name = "Bearer Authentication")
-    public AboutToStartOrSubmitCallbackResponse generateC7ResponseDraftDocument(
-        @RequestHeader(HttpHeaders.AUTHORIZATION) @Parameter(hidden = true) String authorisation,
-        @RequestBody @Parameter(name = "CaseData") uk.gov.hmcts.reform.ccd.client.model.CallbackRequest request
-    ) throws Exception {
-        CaseData caseData = CaseUtils.getCaseData(request.getCaseDetails(), objectMapper);
-
-        Map<String, Object> caseDataUpdated = request.getCaseDetails().getData();
-
-        caseDataUpdated.putAll(documentGenService.generateC7DraftDocuments(authorisation, caseData));
-
-        return AboutToStartOrSubmitCallbackResponse.builder().data(caseDataUpdated).build();
-    }
-
-}
