+++ src/main/java/uk/gov/hmcts/reform/prl/enums/respondentsolicitor/RespondentWelshNeedsListEnum.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.prl.enums.solicitorresponse;
+package uk.gov.hmcts.reform.prl.enums.respondentsolicitor;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
