+++ src/main/java/uk/gov/hmcts/reform/prl/models/complextypes/Behaviours.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.reform.prl.models.complextypes;
 import lombok.Builder;
 import lombok.Data;
 import uk.gov.hmcts.reform.prl.enums.YesOrNo;
-import uk.gov.hmcts.reform.prl.enums.solicitorresponse.AbuseTypes;
+import uk.gov.hmcts.reform.prl.enums.respondentsolicitor.AbuseTypes;
 
 @Builder
 @Data
