+++ src/main/java/uk/gov/hmcts/reform/prl/models/complextypes/solicitorresponse/AttendToCourt.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import uk.gov.hmcts.reform.prl.enums.YesOrNo;
-import uk.gov.hmcts.reform.prl.enums.solicitorresponse.RespondentWelshNeedsListEnum;
+import uk.gov.hmcts.reform.prl.enums.respondentsolicitor.RespondentWelshNeedsListEnum;
 import uk.gov.hmcts.reform.prl.models.Element;
 
 import java.util.List;
